[W3D2] Message를 출력하는 명령을 NavigationMessage 클래스 하나로 묶음
