+++ src/main/java/org/prgrms/kdt/CommandLineApplication.java
@@ -1,11 +1,11 @@
 package org.prgrms.kdt;
 
 import org.prgrms.kdt.command.CommandCreate;
+import org.prgrms.kdt.command.NavigationMessage;
 import org.prgrms.kdt.voucher.Voucher;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
 import java.io.IOException;
-import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -16,15 +16,12 @@ public class CommandLineApplication {
         final var applicationContext = new AnnotationConfigApplicationContext(AppConfiguration.class);
         final Scanner scanner = new Scanner(System.in);
 
-        final List<Optional<Voucher>> mylist = new ArrayList<>();
+        final List<Optional<Voucher>> voucherList = new ArrayList<>();
 
         boolean programRunning = true;
 
         do {
-            System.out.println("=== Voucher Program ===");
-            System.out.println("Type 'exit' for Exit.");
-            System.out.println("Type 'create' to create a new voucher");
-            System.out.println("Type 'list' to list all vouchers");
+            new NavigationMessage();
 
             final String commandInput = scanner.nextLine();
 
@@ -34,19 +31,18 @@ public class CommandLineApplication {
                     break;
 
                 case "create":
-                    CommandCreate.voucherCreateMessage();
-                    mylist.add(CommandCreate.createVoucherType());
+                    NavigationMessage.voucherCreateMessage();
+                    voucherList.add(CommandCreate.createVoucherType());
                     break;
 
                 case "list":
-                    for (final Optional<Voucher> voucher : mylist) {
+                    for (final Optional<Voucher> voucher : voucherList) {
                         System.out.println(voucher);
                     }
                     break;
 
                 default:
-                    System.out.println("=== Input type error ===");
-                    System.out.println(MessageFormat.format("{0}은(는) 지원하지 않는 명령어입니다.", commandInput));
+                    NavigationMessage.inputTypeErrorMessage(commandInput);
                     break;
             }
         } while (programRunning);
