+++ src/main/java/org/prgrms/kdt/command/CommandCreate.java
@@ -1,6 +1,5 @@
 package org.prgrms.kdt.command;
 
-import org.prgrms.kdt.voucher.Validation;
 import org.prgrms.kdt.voucher.Voucher;
 import org.prgrms.kdt.voucher.repository.MemoryVoucherRepository;
 
@@ -8,12 +7,6 @@ import java.util.Optional;
 import java.util.Scanner;
 
 public class CommandCreate {
-    public static void voucherCreateMessage() {
-        System.out.println("=== Voucher Create ===");
-        System.out.println("Choose the type of voucher.");
-        System.out.println("- FixedAmountVoucher");
-        System.out.println("- PercentDiscountVoucher");
-    }
 
     public static Optional<Voucher> createVoucherType() {
         final Scanner scanner = new Scanner(System.in);
@@ -24,13 +17,13 @@ public class CommandCreate {
 
             switch (voucherType) {
                 case "FixedAmountVoucher", "PercentDiscountVoucher" -> {
-                    new HowMuchDiscountMessage(voucherType);
+                    NavigationMessage.howMuchDiscountMessage(voucherType);
                     return Optional.ofNullable(MemoryVoucherRepository.create( // voucher 생성
-                            voucherType, Validation.inputValueValidation(voucherType) // 유효한 값인지 생성 및 검사
+                            voucherType, ValueValidation.inputValueValidation(voucherType) // 유효한 값인지 생성 및 검사
                     ));
                 }
 
-                default -> new HowMuchDiscountMessage(voucherType);
+                default -> NavigationMessage.howMuchDiscountMessage(voucherType);
             }
         }
     }
